backend/requirements.txt
Flask
Flask-CORS
flask-socketio
python-socketio
pymongo
python-dotenv
bcrypt
PyJWT
google-genai
faiss-cpu
sentence-transformers
numpy
scikit-learn
gunicorn
PyPDF2
python-docx
Werkzeug
eventlet